% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{validate_causal_spec}
\alias{validate_causal_spec}
\title{Validate Causal Specification}
\usage{
validate_causal_spec(x)
}
\arguments{
\item{x}{A causal_spec object}
}
\value{
The validated object (invisibly)
}
\description{
Validates a causal_spec object
}
