% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrumental_variables.R
\name{summary.iv_effect}
\alias{summary.iv_effect}
\title{Summary method for iv_effect}
\usage{
\method{summary}{iv_effect}(object, ...)
}
\arguments{
\item{object}{An iv_effect object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for iv_effect
}
